eslint: remove node_modules blacklist
This should be unnecessary as it's the default. It was added in an
attempt to fix the remote Travis CI build, whereas the local build would
complete cleanly.

Ultimately the remote build was fixed by updating target Node.js
versions in a7e380ad13da4f62b2816821e5470e316a97b449, so this should be
safe to remove.